+++ tests/snapshots/types.rs
@@ -1,8 +1,8 @@
 // Every fixed-size type — these all copy freely
 
 fn main() {
-    let small: i16 = 42;
-    let count: i32 = 100000;
+    let small: i32 = 42;
+    let count: i64 = 100000;
     let huge: i64 = 9000000000;
     let pi: f32 = 3.14;
     let ratio: f64 = 2.5;
